Rename file to match token
